Finish proof of spmf_of_bind_pmfsr and measure_bind_pmfsr.
Note: I had to change the definition of measure_bind_pmfsr, this is mainly
because the bind in the Giry_Monad seems to require too strong conditions
to work with, i.e., if given (m >>= f) then f needs to be a subprob measure
even on null-sets of m. If (f x) is not wf_pmfsr than the latter may not be
true.

However, even with the slightly altered definition it is straightforward
to show spmf_of_bind_pmfsr (without modifying the statement).